requirements.txt
fastapi
uvicorn
python-docx
fitz
faiss-cpu
python-dotenv
requests
numpy
google-generativeai
scikit-learn
